feat: make it work (#2)
This PR takes the basic implementation of this app (which I made while I
was still allowed to push to main) and iterates on it in tandem with the
work in https://github.com/grafbase/grafbase/pull/2090 (see that PR for
context)

I've also added some CI that'll build releases. Once this is merged I'll
update the grafbase repo to use the binaries instead of building on
every CI run.